Skip if untyped
I had hoped to avoid this because it technically means that you could
have valid code that Sorbet rejects. But in practice, this could only be
happening if people were already writing signatures with overloads where
picking the correct overload depends on block arity, and that isn't the
case right now so I don't actually think this is a breaking change.
It'll just have to be something that people workaround when Sorbet tells
them that something is `nil` that's not actually `nil`.