Write trees to the cache even if GlobalState did not change
- First attempt: c9418bb
- (Semantically) reverted in: 6f56fa4

This change reimplements the spirit of the first attempt, but **only**
when `!gs.wasModified()`. In particular, we continue to not write trees
in all the other cases. For example, if `gs.hadCriticalError()` or if
`kvstoreChangedSinceGsCreation` we do not attempt to write the trees.

The problem with the first attempt was that we would **always** write
the trees into the cache, which might mention `NameRef`'s which are in
GlobalState. But if `retainGlobalState` had opted **not** to write
`GlobalState` into the cache, then those `NameRef`s would point to
nowhere.

I'm not sure of **all** the reasons why we were seeing
`retainGlobalState` opt to not cache `GlobalState`, but I don't think it
was just because `!gs.wasModified`. In particular, I know that it's
possible for the issue described in #8166 to have caused the cache to be
poisoned as a result of landing the first attempt (I know this because I
modified `test/cli/stripe-packages-cache` and was able to confirm the
cache crashes when a file has the `ForwardedRestArg` node).

I don't know if there are **other** internal errors that are common that
are causing this, but in any case, this PR leaves it like it was
originally, so that neither GlobalState nor trees are cached when
there was an internal error.

Similarly I don't know if it was the case that we have some flow which
routinely overwrites the cache file from another Sorbet process.
Technically someone could have invoked Sorbet twice at the same time
(e.g. in different terminal windows) but I would find it really
surprising if that were the root cause given how widespread the cache
poisoning was.