Show changes to CLI tests
The `types.input.files.kvstore.write : 0` lines are because that line is
written only if `cacheTreesAndFiles` is called, which it might not have
before, if `GlobalState` was not written.

In all other cases, the changes now match what the comments in the
output file say should be happening (e.g. the `... (warm cache)` lines
have no misses, and the `... (one cache miss)` lines are once again
accurate).